Add a prepare playbook that imports the update and python playbooks
This upgrades the base image, installs python3/pip3, and removes
python2/pip2.  This leaves the base image in a state just like it
would be if we were building an AMI via packer, and is a much more
realistic canvas on which to perform our molecule tests.

It may be possible to drop the python playbook once both:
* Python2 and Python3 are no longer co-existing in the distros we use
* Ansible no longer insists on using /usr/bin/python if it exists