Modify code to download a fresh tarball from the CS website
Previously we were using a supposedly blessed tarball that we kept in
an S3 bucket, but I have since learned that there is nothing special
about those tarballs.

There was also an issue where the tarballs we were using were too
outdated to upgrade themselves, as described in
cisagov/ansible-role-cobalt-strike#43.  This change resolves that issue.