Add example code showing how to pass GHA secrets to Molecule
In many repos that inherit from this skeleton we need to specify a
bucket name where third-party installers, packages, licenses, etc. are
stored.  Previously we made this bucket name a role var with a default
value equal to the name of the bucket in our production COOL.  Now
we have multiple COOL environments, and we eventually want to test in
all environments to ensure that the third-party buckets in all
environments contain the expected objects.

Therefore, it no longer makes sense to use a default value for the
role variable.  Instead we lookup an environment variable in
converge.yml and have the GHA workflow set that variable when running
Molecule.

Co-authored-by: David Redmin <david.redmin@gwe.cisa.dhs.gov>